Move ThreadHelper to test package
