+++ src/e2e/java/teammates/e2e/cases/InstructorFeedbackEditPageE2ETest.java
@@ -14,9 +14,9 @@ import teammates.common.datatransfer.questions.FeedbackContributionQuestionDetai
 import teammates.common.datatransfer.questions.FeedbackTextQuestionDetails;
 import teammates.common.util.AppUrl;
 import teammates.common.util.Const;
-import teammates.common.util.ThreadHelper;
 import teammates.e2e.pageobjects.FeedbackSubmitPage;
 import teammates.e2e.pageobjects.InstructorFeedbackEditPage;
+import teammates.test.ThreadHelper;
 
 /**
  * SUT: {@link Const.WebPageURIs#INSTRUCTOR_SESSION_EDIT_PAGE}.
