+++ src/e2e/java/teammates/e2e/pageobjects/InstructorCourseDetailsPage.java
@@ -13,7 +13,7 @@ import org.openqa.selenium.support.FindBy;
 import teammates.common.datatransfer.attributes.CourseAttributes;
 import teammates.common.datatransfer.attributes.InstructorAttributes;
 import teammates.common.datatransfer.attributes.StudentAttributes;
-import teammates.common.util.ThreadHelper;
+import teammates.test.ThreadHelper;
 
 /**
  * Represents the instructor course details page of the website.
