+++ src/e2e/java/teammates/e2e/pageobjects/InstructorStudentListPage.java
@@ -12,7 +12,7 @@ import org.openqa.selenium.WebElement;
 
 import teammates.common.datatransfer.attributes.CourseAttributes;
 import teammates.common.datatransfer.attributes.StudentAttributes;
-import teammates.common.util.ThreadHelper;
+import teammates.test.ThreadHelper;
 
 /**
  * Page Object Model for instructor student list page.
