+++ src/test/java/teammates/ui/webapi/FeedbackSessionClosingRemindersActionTest.java
@@ -10,9 +10,9 @@ import teammates.common.util.Const;
 import teammates.common.util.EmailType;
 import teammates.common.util.EmailWrapper;
 import teammates.common.util.TaskWrapper;
-import teammates.common.util.ThreadHelper;
 import teammates.common.util.TimeHelper;
 import teammates.common.util.TimeHelperExtension;
+import teammates.test.ThreadHelper;
 import teammates.ui.request.SendEmailRequest;
 
 /**
